Conflicts:
	src/test/java/com/bonc/test/CommonTest3.java

+++ src/main/java/com/bonc/SpringBootAdminServerRunner.java
@@ -1,6 +1,8 @@
 package com.bonc;
 
 import java.io.IOException;
+import java.io.InputStreamReader;
+import java.nio.CharBuffer;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -12,6 +14,7 @@ import org.springframework.context.event.ContextClosedEvent;
 import org.springframework.context.event.ContextStoppedEvent;
 import org.springframework.context.event.EventListener;
 import org.springframework.core.annotation.Order;
+import org.springframework.util.StringUtils;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 /**
@@ -20,9 +23,9 @@ import org.springframework.web.bind.annotation.RestController;
  * @date   2019年7月5日上午11:12:40
  * @Description TODO
  */
-@Order
-@RestController
-@ConditionalOnExpression("'${spring.boot.admin.server.start-command:}'!=''")
+//@Order
+//@RestController
+//@ConditionalOnExpression("'${spring.boot.admin.server.start-command:}'!=''")
 public class SpringBootAdminServerRunner implements CommandLineRunner{
 	private Process process;	
 	private final static Logger LOG = LoggerFactory.getLogger(SpringBootAdminServerRunner.class);
@@ -51,14 +54,20 @@ public class SpringBootAdminServerRunner implements CommandLineRunner{
 			return;
 		String[] cmds= {"cmd.exe","/c",pidCommand};
 		Process process=Runtime.getRuntime().exec(cmds);
-		byte[] b=new byte[8*1024];
-		process.getInputStream().read(b);
+		CharBuffer cb=CharBuffer.allocate(8*1024);
+		StringBuilder sb=new StringBuilder();
+		InputStreamReader reader=new InputStreamReader(process.getInputStream());
+		while(reader.read(cb)>-1) {
+			cb.flip();
+			sb.append(cb);
+			cb.clear();
+		};
 		process.destroy();
-		String str=new String(b);
+		String str=sb.toString();
 		String[] strs=str.split("\r\n");
 		for (int i = 0; i < strs.length; i++) {
 			String[] temp=strs[i].split("\\s+");
-			if(!temp[temp.length-1].equals("0")){
+			if(!temp[temp.length-1].equals("0")&&!StringUtils.isEmpty(temp[temp.length-1])){
 				//按pid关闭
 				cmds[2]="taskkill /F /PID "+temp[temp.length-1];
 				process=Runtime.getRuntime().exec(cmds);
@@ -68,10 +77,11 @@ public class SpringBootAdminServerRunner implements CommandLineRunner{
 		}
 	}
 	@RequestMapping("/stopSpringbootAdminServer")
-	public Object stopSpringbootAdminServer(){
+	public Object stopSpringbootAdminServer() throws IOException{
 		String result="Spring boot admin server doesn't run";
 		if(process!=null &&process.isAlive()) {
-			process.destroy();
+//			process.destroy();
+			stopPreProcess();
 			result="Stop spring boot admin server success";
 			LOG.info("Stop spring boot admin server success");
 		}
@@ -87,7 +97,7 @@ public class SpringBootAdminServerRunner implements CommandLineRunner{
 		return result;
 	}
 	@EventListener(classes= {ContextClosedEvent.class,ContextStoppedEvent.class})
-	public void shutdown(ApplicationEvent ae) {
+	public void shutdown(ApplicationEvent ae) throws IOException {
 		System.out.println(ae);
 		stopSpringbootAdminServer();
 	}
